New files, more fun

+++ Numerical_Methods/Hanoi/main.c
@@ -6,7 +6,7 @@
  * Function prototyping.
  * @PARAM set, takes the number set you want to check 
  * */
-uint64_t hanoi(int set);
+__uint128_t hanoi(int set);
 /*
  * @PARAM argc, argument counter
  * @PARAM argv, arguments in a char pointer array
@@ -18,9 +18,9 @@ int main(int argc, char * argv[]) {
     }
     int set = atoi(argv[1]);
     int x = 0;
-    uint64_t hanoi_number = hanoi(set);
+    __uint128_t hanoi_number = hanoi(set);
     char han[] = {hanoi_number};
-    printf("Hanoi nr:\t%d\tSet size:\t%d\n", hanoi_number, set);
+    printf("Hanoi nr:\t%lld\tSet size:\t%d\n", hanoi_number, set);
     hanoi_number = 0;
     while (hanoi_number < 1000000) {
         hanoi_number = hanoi(x);
@@ -36,7 +36,7 @@ int main(int argc, char * argv[]) {
  * @PARAM set, is the number of blocks to move in the tower of hanoi game. 
  * @RETURN uint64_t 2*hanoi(set-1)+1 if set > 0
  * */
-uint64_t hanoi(int set) {
+__uint128_t hanoi(int set) {
     if (set == 0)
         return 0;
     else
